refac: sidebar
